+++ database/migrations/2026_01_29_000005_create_paiements_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreignId('echeance_id')->nullable()->constrained('echeances')->onDelete('set null');
             $table->enum('type_paiement', ['loyer', 'caution', 'avance', 'frais_agence', 'prix_achat', 'arrhes'])->default('loyer');
             $table->enum('statut', ['en_attente', 'paye', 'partiel', 'annule'])->default('en_attente');
-            $table->decimal('montant', 15, 2);
-            $table->decimal('commission_agence', 10, 2)->nullable();
+            $table->unsignedBigInteger('montant');
+            $table->unsignedBigInteger('commission_agence')->nullable();
             $table->enum('type_commission', ['pourcentage', 'montant'])->nullable();
             $table->date('date_paiement');
             $table->enum('methode_paiement', ['espèces', 'virement', 'chèque', 'carte_bancaire', 'mobile_money', 'autre'])->default('virement');
